New 'memorise forgotten' option.
This option can be used if the database is lost. All issues which CappBot has commented on will be logged in the database the way they would have looked if all had gone right. Issues without a CappBot comment are still assumed to be new if they're not in the database.